// src/app/kimios-client-api/model/shareByEmailFullContactParam.ts
/**
 * Kimios API
 * Kimios Sample API
 *
 * OpenAPI spec version: 1.3-SNAPSHOT - 33617c598e5995c9680a7c36de40cd9ae07b1f40
 * Contact: documentation@kimios.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { MailContact } from './mailContact';


export interface ShareByEmailFullContactParam { 
    sessionId?: string;
    documentIds?: Array<number>;
    recipients?: Array<MailContact>;
    subject?: string;
    content?: string;
    senderAddress?: string;
    senderName?: string;
    defaultSender?: boolean;
    password?: string;
    expirationDate?: string;
}